r162: more meaningful traversal over inversions
If the inversion involves more than one segments, the previous version
arbitrarily duplicates one segment. In the "bubble" output, the two sides of
the bubble are not reverse complemented, either. The new version duplicates all
segments in the bubble and prints the desired output.

However, I am not sure if the new change has issues with more complex
topologies (e.g. cycles interleaved with inversions).