r175: deal with another corner case
The last several commits aim to handle undesired linearizations in the presence
of inversions. The older bubble command may generate excessively long events,
caused by vertices not moved out of the AVL tree in time. The current
implementation moves a vertex out of the tree in the following order:

1. A vertex whose complement has been moved out.
2. If there are no vertices without parents, choose a vertex whose complement
   is also in the tree.
3. Otherwise, choose the "smallest" vertex.

The current heuristic works for the 20-genome graph generated by minigraph-0.9.